requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable management
Python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# For hugging face embedding 
sentence-transformers

# UI libraries
streamlit

# Youtube fetching
youtube-transcript-api

# Vector store
faiss-cpu
tiktoken